refactor(expand contract): split out extract/contract from both CSG & CAG, restructured accordingly
