+++ src/core/CAG.js
@@ -1,4 +1,3 @@
-const {EPS, angleEPS} = require('./constants')
 const {Connector} = require('./connectors')
 const Vertex3D = require('./math/Vertex3')
 const Vector2D = require('./math/Vector2')
@@ -6,7 +5,7 @@ const Vector3D = require('./math/Vector3')
 const Polygon = require('./math/Polygon3')
 
 const {fromPolygons} = require('./CSGFactories')
-const {fromSides, fromFakeCSG, fromPoints, fromPointsNoCheck} = require('./CAGFactories')
+const {fromSides, fromFakeCSG} = require('./CAGFactories')
 
 const canonicalize = require('./utils/canonicalize')
 const retesselate = require('./utils/retesellate')
@@ -17,7 +16,7 @@ const {area, getBounds} = require('./utils/cagMeasurements')
 const {overCutInsideCorners} = require('../api/ops-cnc')
 const {extrudeInOrthonormalBasis, extrudeInPlane, extrude, rotateExtrude} = require('../api/ops-extrusions')
 const cagoutlinePaths = require('../api/cagOutlinePaths')
-
+const {expand, contract, expandedShellOfCAG} = require('../api/ops-expandContract')
 /**
  * Class CAG
  * Holds a solid area geometry like CSG but 2D.
@@ -102,100 +101,18 @@ CAG.prototype = {
     return fromSides(newsides)
   },
 
+  // ALIAS !
   expandedShell: function (radius, resolution) {
-    resolution = resolution || 8
-    if (resolution < 4) resolution = 4
-    let cags = []
-    let pointmap = {}
-    let cag = this.canonicalized()
-    cag.sides.map(function (side) {
-      let d = side.vertex1.pos.minus(side.vertex0.pos)
-      let dl = d.length()
-      if (dl > EPS) {
-        d = d.times(1.0 / dl)
-        let normal = d.normal().times(radius)
-        let shellpoints = [
-          side.vertex1.pos.plus(normal),
-          side.vertex1.pos.minus(normal),
-          side.vertex0.pos.minus(normal),
-          side.vertex0.pos.plus(normal)
-        ]
-        //      let newcag = fromPointsNoCheck(shellpoints);
-        let newcag = fromPoints(shellpoints)
-        cags.push(newcag)
-        for (let step = 0; step < 2; step++) {
-          let p1 = (step === 0) ? side.vertex0.pos : side.vertex1.pos
-          let p2 = (step === 0) ? side.vertex1.pos : side.vertex0.pos
-          let tag = p1.x + ' ' + p1.y
-          if (!(tag in pointmap)) {
-            pointmap[tag] = []
-          }
-          pointmap[tag].push({
-            'p1': p1,
-            'p2': p2
-          })
-        }
-      }
-    })
-    for (let tag in pointmap) {
-      let m = pointmap[tag]
-      let angle1, angle2
-      let pcenter = m[0].p1
-      if (m.length === 2) {
-        let end1 = m[0].p2
-        let end2 = m[1].p2
-        angle1 = end1.minus(pcenter).angleDegrees()
-        angle2 = end2.minus(pcenter).angleDegrees()
-        if (angle2 < angle1) angle2 += 360
-        if (angle2 >= (angle1 + 360)) angle2 -= 360
-        if (angle2 < angle1 + 180) {
-          let t = angle2
-          angle2 = angle1 + 360
-          angle1 = t
-        }
-        angle1 += 90
-        angle2 -= 90
-      } else {
-        angle1 = 0
-        angle2 = 360
-      }
-      let fullcircle = (angle2 > angle1 + 359.999)
-      if (fullcircle) {
-        angle1 = 0
-        angle2 = 360
-      }
-      if (angle2 > (angle1 + angleEPS)) {
-        let points = []
-        if (!fullcircle) {
-          points.push(pcenter)
-        }
-        let numsteps = Math.round(resolution * (angle2 - angle1) / 360)
-        if (numsteps < 1) numsteps = 1
-        for (let step = 0; step <= numsteps; step++) {
-          let angle = angle1 + step / numsteps * (angle2 - angle1)
-          if (step === numsteps) angle = angle2 // prevent rounding errors
-          let point = pcenter.plus(Vector2D.fromAngleDegrees(angle).times(radius))
-          if ((!fullcircle) || (step > 0)) {
-            points.push(point)
-          }
-        }
-        let newcag = fromPointsNoCheck(points)
-        cags.push(newcag)
-      }
-    }
-    let result = new CAG()
-    result = result.union(cags)
-    return result
+    return expandedShellOfCAG(this, radius, resolution)
   },
 
+  // ALIAS !
   expand: function (radius, resolution) {
-    let result = this.union(this.expandedShell(radius, resolution))
-    return result
+    return expand(this, radius, resolution)
   },
 
   contract: function (radius, resolution) {
-    let result = this.subtract(this.expandedShell(radius, resolution))
-    return result
+    return contract(this, radius, resolution)
   },
 
   // ALIAS !
